Update and rename permutation and comnbination.c to Permutation and Combination.c
This code provides functions to calculate factorial, permutation, and combination. The factorial function calculates the factorial of a given number. The permutation function calculates the permutation of n and r using the formula n! / (n-r)!. The combination function calculates the combination of n and r using the formula n! / (r! * (n-r)!).

In the main function, the user is prompted to enter the values of n and r, and then the permutation and combination are calculated and displayed.